Split into data_decoded.rs and page.rs
