+++ src/routes/chains/handlers.rs
@@ -1,6 +1,6 @@
 use crate::cache::cache_operations::RequestCached;
 use crate::common::models::backend::chains::ChainInfo as BackendChainInfo;
-use crate::common::models::Page;
+use crate::common::models::page::Page;
 use crate::config::{chain_info_cache_duration, chain_info_request_timeout};
 use crate::providers::info::{DefaultInfoProvider, InfoProvider};
 use crate::routes::chains::models::ChainInfo as ServiceChainInfo;
