+++ src/routes/transactions/converters/tests/details.rs
@@ -1,8 +1,8 @@
 use crate::common::models::addresses::AddressEx;
 use crate::common::models::backend::transactions::{ModuleTransaction, MultisigTransaction};
 use crate::common::models::backend::transfers::Transfer as TransferDto;
-use crate::common::models::ParamValue::SingleValue;
-use crate::common::models::{DataDecoded, Operation, Parameter};
+use crate::common::models::data_decoded::ParamValue::SingleValue;
+use crate::common::models::data_decoded::{DataDecoded, Operation, Parameter};
 use crate::providers::info::*;
 use crate::routes::transactions::models::details::{
     DetailedExecutionInfo, ModuleExecutionDetails, MultisigConfirmation, MultisigExecutionDetails,
