+++ src/routes/transactions/handlers/details.rs
@@ -3,7 +3,7 @@ extern crate reqwest;
 use crate::cache::cache_operations::RequestCached;
 use crate::common::models::backend::transactions::{ModuleTransaction, MultisigTransaction};
 use crate::common::models::backend::transfers::Transfer;
-use crate::common::models::Page;
+use crate::common::models::page::Page;
 use crate::config::transaction_request_timeout;
 use crate::providers::info::{DefaultInfoProvider, InfoProvider};
 use crate::routes::transactions::models::details::TransactionDetails;
