+++ src/routes/transactions/handlers/history.rs
@@ -2,7 +2,7 @@ extern crate reqwest;
 
 use crate::cache::cache_operations::RequestCached;
 use crate::common::models::backend::transactions::{CreationTransaction, Transaction};
-use crate::common::models::{Page, PageMetadata};
+use crate::common::models::page::{Page, PageMetadata};
 use crate::config::transaction_request_timeout;
 use crate::providers::info::{DefaultInfoProvider, InfoProvider};
 use crate::routes::transactions::handlers::offset_page_meta;
