+++ src/utils/tests/method_names.rs
@@ -1,4 +1,4 @@
-use crate::common::models::DataDecoded;
+use crate::common::models::data_decoded::DataDecoded;
 use crate::utils::{ERC20_TRANSFER_METHODS, ERC721_TRANSFER_METHODS, SETTINGS_CHANGE_METHODS};
 
 #[test]
